+++ packages/docs/site/CHANGELOG.md
@@ -10,6 +10,7 @@
 		- as of now dynamic slots aren't supported so we can't document named slots
 		- it supports 1 level of depth only so slots being components with slots
 			themselves won't work
+- a11y menu: added a `z-index` to the menu container
 
 ## `@svizzle/site` v0.4.0